Dashboard component finished
