+++ src/App.js
@@ -3,6 +3,7 @@ import { Routes, Route } from "react-router-dom";
 import Dashboard from "./components/dashboard/Dashboard";
 import NewEvent from "./components/newEvent/NewEvent";
 import { EditEvent } from "./components/eventEdit";
+import DetailEvent from "./components/eventClicked/DetailEvent";
 import Login from "./components/Login";
 import Profile from "./components/Profile";
 import "./styles/app.scss";
@@ -15,7 +16,8 @@ function App() {
         <Route path="/dashboard" element={<Dashboard />} />
         <Route path="/createEvent" element={<NewEvent />} />
         <Route path="/editEvent" element={<EditEvent />} />
-        <Route path="/profile" element={<EditEvent />} />
+        <Route path="/detailEvent" element={<DetailEvent />} />
+        <Route path="/profile" element={<Profile />} />
       </Routes>
     </div>
   );
