+++ src/Components/modal/Modal.js
@@ -1,0 +1,38 @@
+import { useState } from "react";
+import { Navigate } from "react-router-dom";
+
+import "./modalStyle.scss";
+
+const Modal = () => {
+  const [goToProfile, setGoToProfile] = useState(false);
+  const [signOut, setSignOut] = useState(false);
+
+  if (signOut) {
+    return <Navigate to="/" />;
+  } else if (goToProfile) {
+    return <Navigate to="/profile" />;
+  } else {
+    return (
+      <nav className="profile-sign-out-container">
+        <span
+          className="profile-link"
+          onClick={() => {
+            setGoToProfile(true);
+          }}
+        >
+          View Profile
+        </span>
+        <span
+          className="sign-out-button"
+          onClick={() => {
+            setSignOut(true);
+          }}
+        >
+          Sign Out
+        </span>
+      </nav>
+    );
+  }
+};
+
+export default Modal;
